requirements.txt
absl-py==0.9.0
Adafruit-Blinka==5.2.0
adafruit-circuitpython-busdevice==4.3.1
adafruit-circuitpython-mlx90614==1.2.1
Adafruit-PlatformDetect==2.14.1
Adafruit-PureIO==1.1.5
astor==0.8.1
bluepy==1.3.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cPython==0.0.5
cryptography==2.9.2
dlib==19.20.0
dnspython==1.16.0
dominate==2.5.1
eventlet==0.25.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-MQTT==1.1.1
Flask-SocketIO==4.3.1
future==0.18.2
gast==0.2.2
gevent==20.6.2
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.19.0
oauthlib==3.1.0
opencv-contrib-python==4.1.0.25
opencv-python==4.1.1.26
opt-einsum==3.3.0
paho-mqtt==1.5.0
picamera==1.13
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pybind11==2.5.0
pycparser==2.20
pyftdi==0.51.2
pymongo==3.10.1
pyserial==3.2.1
python-engineio==3.13.1
python-socketio==4.6.0
pyusb==1.0.2
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
rsa==4.6
scipy==1.4.1
serial==0.0.90
six==1.15.0
SpeechRecognition==3.8.1
sysv-ipc==1.0.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
